Revert back to old patching location, as I can't get the new location to work
