+++ sdk/vision/azure-ai-vision-imageanalysis/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0b2 (Unreleased)
+## 1.0.0b2 (2024-02-09)
 
 ### Breaking Changes
